Feat: 'Creating Labels and Annotations' Exercise from Python for Data Science Course
In this commit, I have added the solutions for the 'Creating Labels and Annotations' exercise from the 'Python for Data Science: Essential Training Part 1' course on LinkedIn Learning. The exercise involved creating and customizing labels and annotations for data visualizations using Python. These elements enhance the interpretability and clarity of the plots. This commit demonstrates my understanding and application of these data visualization techniques.