Feat:  'Lemmatizing and Analyzing Textual Data' Exercise from Python for Data Science Course
In this commit, I have added the solutions or notes for the 'Lemmatizing and Analyzing Textual Data' exercise from the 'Python for Data Science: Essential Training Part 1' course on LinkedIn Learning. The exercise involved implementing lemmatization techniques in Python to convert words into their base or dictionary form and performing analysis on the lemmatized textual data. This commit demonstrates my understanding and application of lemmatization for text preprocessing and analysis in NLP tasks.