Feat: 'Introduction to Plotly' Exercise from Python for Data Science Course
In this commit, I have added the solutions or notes for the 'Introduction to Plotly' exercise from the 'Python for Data Science: Essential Training Part 1' course on LinkedIn Learning. The exercise involved learning the basics of Plotly, a powerful data visualization library in Python, and creating interactive plots and charts. This commit demonstrates my understanding and application of Plotly for creating visually appealing and interactive visualizations in data science projects.