Feat:  'Bar Charts and Pie Charts in Plotly' Exercise from Python for Data Science Course
In this commit, I have added the solutions or notes for the 'Bar Charts and Pie Charts in Plotly' exercise from the 'Python for Data Science: Essential Training Part 1' course on LinkedIn Learning. The exercise involved learning how to create bar charts and pie charts using Plotly, a powerful data visualization library in Python. Bar charts are useful for comparing categorical data or displaying frequencies, while pie charts are used to represent proportions or percentages in a dataset. This commit demonstrates my understanding and application of bar charting and pie charting techniques in Plotly for visualizing data in Python.